Near the final stages

+++ exts/executor.py
@@ -13,27 +13,24 @@ class SignalExecutor:
         self.admin_id = config.ADMIN_TG_ID
         self.signal = signal
         self.mt = mt
-        print('Executor init')
+        self.execution_channels = [1, 3, 4, 5]
 
     async def start_execution(self):
-        try:
-            await self.ask_for_confirmation()
-        except asyncio.TimeoutError:
-            await self.execute_signal()
+        await self.ask_for_confirmation()
 
     async def execute_signal(self):
         if self.signal.signal_type == SignalTypes.BUY:
             print('Executing buy order ...')
-            await self.mt.createBuyOrder(self.signal.symbol,
-                                         self.signal.stop_loss,
-                                         self.signal.take_profit,
-                                         self.signal.price)
+            await self.mt.create_buy_order(symbol=self.signal.symbol,
+                                           stop_loss=self.signal.stop_loss,
+                                           take_profit=self.signal.take_profit,
+                                           open_price=self.signal.price)
         elif self.signal.signal_type == SignalTypes.SELL:
             print('Executing sell order ...')
-            await self.mt.createSellOrder(self.signal.symbol,
-                                          self.signal.stop_loss,
-                                          self.signal.take_profit,
-                                          self.signal.price)
+            await self.mt.create_sell_order(symbol=self.signal.symbol,
+                                            stop_loss=self.signal.stop_loss,
+                                            take_profit=self.signal.take_profit,
+                                            open_price=self.signal.price)
 
     async def ask_for_confirmation(self):
         async with self.bot.conversation(self.admin_id) as self.conv:
@@ -46,10 +43,21 @@ Price : {self.signal.price}\nStop Loss : \
 {self.signal.stop_loss}\nTake Profit : {self.signal.take_profit}'
 
             q = await self.conv.send_message(message, buttons=buttons)
-            r = await self.conv.wait_event(events.CallbackQuery)
-            await q.delete()
-            if r.query.data == b'execute':
-                await self.execute_signal()
-            elif r.query.data == b'not_execute':
-                self.conv.cancel()
-                return
+            try:
+                r = await self.conv.wait_event(events.CallbackQuery)
+                await q.delete()
+                if r.query.data == b'execute':
+                    await self.execute_signal()
+                elif r.query.data == b'not_execute':
+                    self.conv.cancel()
+                    return
+            except asyncio.TimeoutError:
+                if self.signal.channel_number in self.execution_channels:
+                    await self.execute_signal()
+                    message = message + '\n\n**Auto-Executed**'
+                    await q.delete()
+                    await self.conv.send_message(message)
+                    return
+                message = message + '\n\n**Terminated**'
+                await q.delete()
+                await self.conv.send_message(message)
